+++ CHANGELOG.md
@@ -17,7 +17,24 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Breaks
 
 
-## 0.1.0 (2024-09-15)
+## 0.1.0 (2025-02-08)
+---
+
+### New
+Add LLM selection and configuration using environment variables, and including OpenAI, Google, Ollama, Anthropic, Hugging Face, Groq, NVIDIA, X AI, Together AI, AI/ML API and OpenRouter [GS-168].
+Add the ideation task to generate ideas for the "AIstronauts-Space Agents on a mission hackathon" from lablab.ai [GS-55].
+Add different environment variables for coding, reasoning and management agents [GS-168].
+Add "openlit" monitoring tool [GS-168].
+Add generate output files on each task [GS-168].
+
+### Changes
+Change: remove the genericuiste dependency.
+
+### Fixes
+Fix the agents and task prompts to effectively work as a team [GS-168].
+
+
+## 0.0.1 (2024-09-15)
 ---
 
 ### New